requirements.txt
pyyaml==6.0.3
playwright==1.55.0
beautifulsoup4==4.13.5
pillow==11.3.0
jinja2==3.1.6
